7 warnings fixed (about int type format)

+++ CPPAssignement/CPPAssignement/functionSET.cpp
@@ -28,8 +28,8 @@ double Set1::f0A(double x, double t)
 
 Set1::Set1(double n)
 {
-	this->size =int(n);
-	this->array = Matrix(n+1);
+	this->size = static_cast<int>(n);
+	this->array = Matrix(static_cast<int>(n+1));
 }
 
 double Set1::getElem(int x, int y)
